requirements.txt
pytest
pyyaml
beautifulsoup4
ruamel.yaml
